Move WidgetMaker functionality into /app
