+++ app/widget_maker.go
@@ -1,4 +1,4 @@
-package maker
+package app
 
 import (
 	"github.com/olebedev/config"
